requirements.txt
discord.py==2.3.2
gspread==5.10.0
oauth2client==4.1.3
python-dotenv==1.0.0
requests==2.31.0
